refactor : split useSocketConnection hooky from KakaoMap component
